LayoutTests/fast/dom/HTMLImageElement/image-lowsrc-getset.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<img id="image" lowsrc="http://hostname/lowsrc.png" style="visibility: hidden;">
<script>
description("This test checks that the lowsrc attribute of img tags are accessible via JavaScript.");
var image = document.getElementById('image');
shouldBe("image.lowsrc", "'http://hostname/lowsrc.png'");
image.lowsrc = "http://hostname/newlowsrc.png";
shouldBe("image.lowsrc", "'http://hostname/newlowsrc.png'");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
